+++ osu.Game.Rulesets.Mania/Beatmaps/ManiaBeatmap.cs
@@ -28,7 +28,6 @@ namespace osu.Game.Rulesets.Mania.Beatmaps
         public ManiaBeatmap(StageDefinition defaultStage)
         {
             Stages.Add(defaultStage);
-            BeatmapInfo.Ruleset = new ManiaRuleset().RulesetInfo;
         }
     }
 }
